+++ Sources/Vec2D.swift
@@ -39,7 +39,7 @@ public func abs(_ vec: Vec2D) -> Double {
 extension Vec2D {
 
     public func rotate(_ degree: Double) -> Vec2D {
-        let radian = Double(Degree(CGFloat(degree)).radian)
+        let radian = Double(Degree(degree).radian)
         return Vec2D(x * cos(radian) - y * sin(radian),
                      x * sin(radian) + y * cos(radian))
     }
